# [Additions] EOS User verification
- Added classes for EOS (my own classes, similar to how Steam has its own classes)
- Added initialization for EOS SDK, it can get the user (and remember it)
- Prepared functionality for EOS achievements/stats but hasn't been tested thoroughly just yet

+++ Plugins/AchievementPlugin/Source/AchievementPlugin/Public/AchievementPlatforms.h
@@ -25,7 +25,7 @@ public:
 	bool InitializePlatform(const EAchievementPlatforms platform);
 	static void ShutdownPlatform();
 
-	static bool SetPlatformAchievementProgress(const FAchievementPlatformData& platformData, int32 progress, bool unlocked);
+	static bool SetPlatformAchievementProgress(const FAchievementPlatformData& platformData, float progress, bool unlocked);
 	static bool PlatformDeleteAchievementProgress(const FAchievementPlatformData& platformData);
 	static bool PlatformDeleteAllAchievementProgress();
